+++ hail/hail/src/is/hail/io/plink/LoadPlink.scala
@@ -30,7 +30,8 @@ case class FamFileConfig(
 )
 
 object LoadPlink {
-  def expectedBedSize(nSamples: Int, nVariants: Long): Long = 3 + nVariants * ((nSamples + 3) / 4)
+  def expectedBedSize(nSamples: Int, nVariants: Int): Long =
+    3 + nVariants.toLong * ((nSamples + 3) / 4)
 
   def parseBim(
     ctx: ExecuteContext,
@@ -317,8 +318,15 @@ object MatrixPLINKReader {
     )
     assert(locusAllelesType == fullMatrixType.rowKeyStruct)
 
-    new MatrixPLINKReader(params, referenceGenome, fullMatrixType, nVariants, sampleInfo, contexts,
-      partitioner)
+    new MatrixPLINKReader(
+      params,
+      referenceGenome,
+      fullMatrixType,
+      nVariants.toLong,
+      sampleInfo,
+      contexts,
+      partitioner,
+    )
   }
 }
 
@@ -530,7 +538,7 @@ class MatrixPLINKReader(
             }
 
             if (hasRowUID)
-              rvb.addLong(i)
+              rvb.addLong(i.toLong)
 
             rvb.endStruct()
