+++ hail/hail/src/is/hail/types/virtual/BlockMatrixType.scala
@@ -400,8 +400,8 @@ case class BlockMatrixType(
   }
 
   def blockShape(i: Int, j: Int): (Long, Long) = {
-    val r = if (i == nRowBlocks - 1) nRows - (i * blockSize) else blockSize
-    val c = if (j == nColBlocks - 1) nCols - (j * blockSize) else blockSize
+    val r = if (i == nRowBlocks - 1) nRows - (i * blockSize) else blockSize.toLong
+    val c = if (j == nColBlocks - 1) nCols - (j * blockSize) else blockSize.toLong
     r -> c
   }
